[IMP] account, product, purchase, _*: refine catalog sections
_*= purchase_stock, sale

Follow-up to [previous commit](https://github.com/odoo/odoo/commit/583456a8f12f1680e1ca95fbb1dc833a1b33bc64) that introduced sections in the
catalog.

This commit refactors and simplifies the catalog sections logic, and relocates
it to the *account* module, as the concept of sections properly belongs there.

It also introduces a small UX improvement by adding a grab cursor icon when
reordering sections in the catalog.

Additionally, it addresses several issues:

- Products were not being correctly displayed under the default section when
  navigating from a order to the catalog or when reloading the catalog page.
  This was caused by the kanban model being initialized before the search panel,
  leaving `section_id` unset and resulting in `None` being passed in RPC call.

- The text 'All' used to appear at the top of the sidebar when no filters
  (e.g. category or tags) were available. Since sections are now always present,
  this label is no longer required.

- When no filters were available, the sidebar collapsed after each interaction.
  With sections now in place, this behavior is no longer desirable and has been
  removed.

task-4593041
