+++ addons/account/static/src/components/product_catalog/kanban_record.js
@@ -1,13 +1,30 @@
+import { useSubEnv } from "@odoo/owl";
 import { ProductCatalogKanbanRecord } from "@product/product_catalog/kanban_record";
 import { ProductCatalogAccountMoveLine } from "./account_move_line";
 import { patch } from "@web/core/utils/patch";
 
 patch(ProductCatalogKanbanRecord.prototype, {
+    setup() {
+        super.setup();
+
+        useSubEnv({
+            ...this.env,
+            selectedSectionId: this.env.searchModel.selectedSection?.sectionId,
+        });
+    },
+
     get orderLineComponent() {
         if (this.env.orderResModel === "account.move") {
             return ProductCatalogAccountMoveLine;
         }
         return super.orderLineComponent;
+    },
+
+    _getUpdateQuantityAndGetPriceParams() {
+        return {
+            ...super._getUpdateQuantityAndGetPriceParams(),
+            section_id: this.env.selectedSectionId,
+        };
     },
 
     addProduct(qty = 1) {
@@ -16,4 +33,20 @@ patch(ProductCatalogKanbanRecord.prototype, {
         }
         super.addProduct(qty);
     },
-});
+
+    updateQuantity(quantity) {
+        const lineCountChange = (quantity > 0) - (this.productCatalogData.quantity > 0);
+        if (lineCountChange !== 0) {
+            this.notifyLineCountChange(lineCountChange);
+        }
+
+        super.updateQuantity(quantity);
+    },
+
+    notifyLineCountChange(lineCountChange) {
+        this.env.searchModel.trigger('section-line-count-change', {
+            sectionId: this.env.selectedSectionId,
+            lineCountChange: lineCountChange,
+        });
+    },
+})
