+++ addons/account/static/src/components/product_catalog/kanban_view.js
@@ -1,0 +1,9 @@
+import { productCatalogKanbanView } from "@product/product_catalog/kanban_view";
+import { patch } from "@web/core/utils/patch";
+import { AccountProductCatalogSearchModel } from "./search/search_model";
+import { AccountProductCatalogSearchPanel} from "./search/search_panel";
+
+patch(productCatalogKanbanView, {
+    SearchModel: AccountProductCatalogSearchModel,
+    SearchPanel: AccountProductCatalogSearchPanel,
+});
