+++ addons/account/static/src/components/product_catalog/search/search_panel.scss
@@ -12,3 +12,11 @@
         font-weight: bold;
     }
 }
+
+.o_row_handle {
+    @include o-grab-cursor;
+    color: #adb5bd;
+    &:hover {
+        color: #666666;
+    }
+}
