+++ addons/account/tests/test_tour.py
@@ -108,8 +108,7 @@ class TestUi(AccountTestInvoicingHttpCommon):
         self.assertTrue(self.env.user.has_group('account.group_partial_purchase_deductibility'))
         self.start_tour("/odoo/vendor-bills/new", 'deductible_amount_column', login=self.env.user.login)
 
-    def _test_add_section_from_product_catalog_on_invoice_tour(self):
-        # TODO SHRM fix tour and uncomment
+    def test_add_section_from_product_catalog_on_invoice_tour(self):
         self.product.write({'is_favorite': True})
         self.start_tour(
             '/odoo/customer-invoices/new',
