+++ addons/product/models/product_catalog_mixin.py
@@ -149,175 +149,3 @@ class ProductCatalogMixin(models.AbstractModel):
         :rtype: float
         """
         return 0
-
-    def _create_order_section(self, child_field, name, position, **kwargs):
-        """ Create a new section in order.
-
-        :param str child_field: Field name of the order's lines (e.g., 'order_line').
-        :param str name: The name of the section to create.
-        :param str position: The position of the section where it should be created, either 'top'
-                              or 'bottom'.
-        :param dict kwargs: Additional values given for inherited models.
-        :return: A dictionary with newly created section's 'id' and 'sequence'.
-        :rtype: dict
-        """
-        line_model, parent_field = self._get_section_model_info()
-
-        if not (line_model and parent_field):
-            return {}
-
-        lines = self[child_field]
-        sequence = 10
-        if lines:
-            sequence = (
-                lines[0].sequence - 1 if position == 'top'
-                else lines[-1].sequence + 1
-            )
-
-        section = self.env[line_model].create({
-            parent_field: self.id,
-            'name': name,
-            'display_type': 'line_section',
-            'sequence': sequence,
-            **kwargs,
-        })
-
-        return {
-            'id': section.id,
-            'sequence': section.sequence,
-        }
-
-    def _get_new_line_sequence(self, child_field, selected_section_id):
-        """ Compute the sequence number for inserting a new line into the order.
-
-        :param str child_field: Field name of the order's lines (e.g., 'order_line').
-        :param int selected_section_id: ID of the section line to insert after.
-        :return: Computed sequence number.
-        :rtype: int
-        """
-        lines = self[child_field]
-        section_lines = lines.filtered_domain([
-            ('display_type', '=', 'line_section'),
-        ]).sorted('sequence')
-
-        if selected_section_id:
-            # Insert after the selected section line
-            sequence = section_lines.filtered_domain([
-                ('id', '=', selected_section_id),
-            ]).sequence + 1
-        elif section_lines:
-            # Insert before the first section (top of the order)
-            sequence = section_lines[0].sequence
-        else:
-            # No sections exist, insert at the end
-            sequence = (lines and lines[-1].sequence + 1) or 10
-
-        for line in lines.filtered_domain([('sequence', '>=', sequence)]):
-            line.sequence += 1
-
-        return sequence
-
-    def _get_order_sections(self, child_field, **kwargs):
-        """ Return section data for the product catalog display.
-
-        :param str child_field: Field name of the order's lines (e.g., 'order_line').
-        :param dict kwargs: Additional values given for inherited models.
-        :return: List of section dicts with 'id', 'name', 'sequence', and 'line_count'.
-        :rtype: list
-        """
-        if not child_field:
-            return []
-        sections = {}
-        no_section_count = 0
-        lines = self[child_field]
-        for line in lines.sorted('sequence'):
-            if line.display_type == 'line_section':
-                sections[line.id] = {
-                    'id': line.id,
-                    'name': line.name,
-                    'sequence': line.sequence,
-                    'line_count': 0,
-                }
-            elif self._is_line_valid_for_section_line_count(line):
-                sec_id = line.get_parent_section_line().id
-                if sec_id and sec_id in sections:
-                    sections[sec_id]['line_count'] += 1
-                else:
-                    no_section_count += 1
-
-        if no_section_count > 0 or not lines:
-            sections[False] = {
-                'id': False,
-                'name': self.env._("No Section"),
-                'sequence': lines[0].sequence - 1 if lines else 0,
-                'line_count': no_section_count,
-            }
-
-        return sorted(sections.values(), key=lambda x: x['sequence'])
-
-    def _get_section_model_info(self):
-        """ Return the model name and parent field for the order lines.
-
-        :return: line_model, parent_field
-        """
-        return False, False
-
-    def _is_line_valid_for_section_line_count(self, line):
-        """ Check if a line is valid for inclusion in the section's line count.
-
-        :param recordset line: A record of an order line.
-        :return: True if this line is a valid, else False.
-        :rtype: bool
-        """
-        return (
-            not line.display_type
-            and line.product_type != 'combo'
-            and line.product_uom_qty > 0
-        )
-
-    def _resequence_order_sections(self, sections, child_field, **kwargs):
-        """ Resequence the sections of the order based on the provided move and target sections.
-        The sections are reordered by updating their sequence numbers.
-
-        :param list sections: A list of dictionaries containing move and target sections.
-        :param str child_field: Field name of the order's lines (e.g., 'order_line').
-        :param dict kwargs: Additional values given for inherited models.
-        :return: A dictonary containing the new sequences of all the sections of order.
-        :rtype: dict
-        """
-        if not child_field:
-            return {}
-        lines = self[child_field].sorted('sequence')
-        move_section, target_section = sections
-
-        move_block = lines.filtered(
-            lambda line: line.id == move_section['id']
-            or line.parent_id.id == move_section['id'],
-        )
-
-        target_block = lines.filtered(
-            lambda line: line.id == target_section['id']
-            or line.parent_id.id == target_section['id'],
-        )
-
-        remaining_lines = lines - move_block
-        insert_after = move_section['sequence'] < target_section['sequence']
-        insert_index = len(remaining_lines)
-        for idx, line in enumerate(remaining_lines):
-            if line.id == (target_block[-1].id if insert_after else target_section['id']):
-                insert_index = idx + 1 if insert_after else idx
-                break
-
-        reordered_lines = (
-            remaining_lines[:insert_index] +
-            move_block +
-            remaining_lines[insert_index:]
-        )
-
-        sections = {}
-        for sequence, line in enumerate(reordered_lines, start=1):
-            line.sequence = sequence
-            if line.display_type == 'line_section':
-                sections[line.id] = sequence
-
-        return sections
