+++ addons/product/static/src/product_catalog/kanban_model.js
@@ -19,18 +19,7 @@ export class ProductCatalogKanbanModel extends RelationalModel {
     async _loadData(params) {
         // if orm have isSample field and its value set to be true then we have sample data as there is no product found for selected vendor, show sample data
         const isSample = this.orm.isSample !== undefined ? this.orm.isSample : false;
-        const selectedSection = this.env.searchModel.selectedSection;
-        if (selectedSection.filtered) {
-            params = {
-                ...params,
-                domain: [...(params.domain || []), ['is_in_selected_section_of_order', '=', true]],
-                context: {
-                    ...params.context,
-                    selected_section_id: selectedSection.sectionId,
-                },
-            };
-        }
-        const result = await super._loadData(params);
+        const result = await super._loadData(...arguments);
         if (!params.isMonoRecord) {
             let records;
             if (params.groupBy?.length) {
@@ -69,7 +58,6 @@ export class ProductCatalogKanbanModel extends RelationalModel {
             product_ids: productIds,
             res_model: params.context.product_catalog_order_model,
             child_field: params.context.child_field,
-            selected_section_id: this.env.searchModel.selectedSection.sectionId,
         }
     }
